Move StorageTest to kinto.core.storage.testing
